Move to top-level use of cluster.Credentials
