+++ cluster/service_database_test.go
@@ -16,11 +16,6 @@ import (
 const shortWait = 1 * time.Second
 const longWait = 5 * time.Second
 
-const (
-	NO_USERNAME = ""
-	NO_PASSWORD = ""
-)
-
 func Test_ServiceExecute(t *testing.T) {
 	ln, mux := mustNewMux()
 	go mux.Serve()
@@ -45,7 +40,7 @@ func Test_ServiceExecute(t *testing.T) {
 		}
 		return nil, errors.New("execute failed")
 	}
-	_, err := c.Execute(executeRequestFromString("some SQL"), s.Addr(), NO_USERNAME, NO_PASSWORD, longWait)
+	_, err := c.Execute(executeRequestFromString("some SQL"), s.Addr(), nil, longWait)
 	if err == nil {
 		t.Fatalf("client failed to report error")
 	}
@@ -63,7 +58,7 @@ func Test_ServiceExecute(t *testing.T) {
 		}
 		return []*command.ExecuteResult{result}, nil
 	}
-	res, err := c.Execute(executeRequestFromString("some SQL"), s.Addr(), NO_USERNAME, NO_PASSWORD, longWait)
+	res, err := c.Execute(executeRequestFromString("some SQL"), s.Addr(), nil, longWait)
 	if err != nil {
 		t.Fatalf("failed to execute query: %s", err.Error())
 	}
@@ -80,7 +75,7 @@ func Test_ServiceExecute(t *testing.T) {
 		}
 		return []*command.ExecuteResult{result}, nil
 	}
-	res, err = c.Execute(executeRequestFromString("some SQL"), s.Addr(), NO_USERNAME, NO_PASSWORD, longWait)
+	res, err = c.Execute(executeRequestFromString("some SQL"), s.Addr(), nil, longWait)
 	if err != nil {
 		t.Fatalf("failed to execute: %s", err.Error())
 	}
@@ -92,7 +87,7 @@ func Test_ServiceExecute(t *testing.T) {
 		time.Sleep(longWait)
 		return nil, nil
 	}
-	_, err = c.Execute(executeRequestFromString("some SQL"), s.Addr(), "", "", shortWait)
+	_, err = c.Execute(executeRequestFromString("some SQL"), s.Addr(), nil, shortWait)
 	if err == nil {
 		t.Fatalf("failed to receive expected error")
 	}
@@ -133,7 +128,7 @@ func Test_ServiceQuery(t *testing.T) {
 		}
 		return nil, errors.New("query failed")
 	}
-	_, err := c.Query(queryRequestFromString("SELECT * FROM foo"), s.Addr(), NO_USERNAME, NO_PASSWORD, longWait)
+	_, err := c.Query(queryRequestFromString("SELECT * FROM foo"), s.Addr(), nil, longWait)
 	if err == nil {
 		t.Fatalf("client failed to report error")
 	}
@@ -151,7 +146,7 @@ func Test_ServiceQuery(t *testing.T) {
 		}
 		return []*command.QueryRows{rows}, nil
 	}
-	res, err := c.Query(queryRequestFromString("SELECT * FROM foo"), s.Addr(), NO_USERNAME, NO_PASSWORD, longWait)
+	res, err := c.Query(queryRequestFromString("SELECT * FROM foo"), s.Addr(), nil, longWait)
 	if err != nil {
 		t.Fatalf("failed to query: %s", err.Error())
 	}
@@ -168,7 +163,7 @@ func Test_ServiceQuery(t *testing.T) {
 		}
 		return []*command.QueryRows{rows}, nil
 	}
-	res, err = c.Query(queryRequestFromString("SELECT * FROM foo"), s.Addr(), NO_USERNAME, NO_PASSWORD, longWait)
+	res, err = c.Query(queryRequestFromString("SELECT * FROM foo"), s.Addr(), nil, longWait)
 	if err != nil {
 		t.Fatalf("failed to query: %s", err.Error())
 	}
@@ -180,7 +175,7 @@ func Test_ServiceQuery(t *testing.T) {
 		time.Sleep(longWait)
 		return nil, nil
 	}
-	_, err = c.Query(queryRequestFromString("some SQL"), s.Addr(), NO_USERNAME, NO_PASSWORD, shortWait)
+	_, err = c.Query(queryRequestFromString("some SQL"), s.Addr(), nil, shortWait)
 	if err == nil {
 		t.Fatalf("failed to receive expected error")
 	}
@@ -242,7 +237,7 @@ func Test_ServiceQueryLarge(t *testing.T) {
 		}
 		return []*command.QueryRows{rows}, nil
 	}
-	res, err := c.Query(queryRequestFromString("SELECT * FROM foo"), s.Addr(), NO_USERNAME, NO_PASSWORD, longWait)
+	res, err := c.Query(queryRequestFromString("SELECT * FROM foo"), s.Addr(), nil, longWait)
 	if err != nil {
 		t.Fatalf("failed to query: %s", err.Error())
 	}
