+++ cluster/service_test.go
@@ -193,12 +193,12 @@ func Test_NewServiceTestExecuteQueryAuthNoCredentials(t *testing.T) {
 		t.Fatalf("failed to set cluster client local parameters: %s", err)
 	}
 	er := &command.ExecuteRequest{}
-	_, err := cl.Execute(er, s.Addr(), NO_USERNAME, NO_PASSWORD, 5*time.Second)
+	_, err := cl.Execute(er, s.Addr(), nil, 5*time.Second)
 	if err != nil {
 		t.Fatal(err)
 	}
 	qr := &command.QueryRequest{}
-	_, err = cl.Query(qr, s.Addr(), NO_USERNAME, NO_PASSWORD, 5*time.Second)
+	_, err = cl.Query(qr, s.Addr(), nil, 5*time.Second)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -235,21 +235,21 @@ func Test_NewServiceTestExecuteQueryAuth(t *testing.T) {
 		t.Fatalf("failed to set cluster client local parameters: %s", err)
 	}
 	er := &command.ExecuteRequest{}
-	_, err := cl.Execute(er, s.Addr(), "alice", "secret1", 5*time.Second)
+	_, err := cl.Execute(er, s.Addr(), makeCredentials("alice", "secret1"), 5*time.Second)
 	if err != nil {
 		t.Fatal("alice improperly unauthorized to execute")
 	}
-	_, err = cl.Execute(er, s.Addr(), "bob", "secret1", 5*time.Second)
+	_, err = cl.Execute(er, s.Addr(), makeCredentials("bob", "secret1"), 5*time.Second)
 	if err == nil {
 		t.Fatal("bob improperly authorized to execute")
 	}
 	qr := &command.QueryRequest{}
-	_, err = cl.Query(qr, s.Addr(), "bob", "secret1", 5*time.Second)
+	_, err = cl.Query(qr, s.Addr(), makeCredentials("bob", "secret1"), 5*time.Second)
 	if err != nil && err.Error() != "unauthorized" {
 		fmt.Println(err)
 		t.Fatal("bob improperly unauthorized to query")
 	}
-	_, err = cl.Query(qr, s.Addr(), "alice", "secret1", 5*time.Second)
+	_, err = cl.Query(qr, s.Addr(), makeCredentials("alice", "secret1"), 5*time.Second)
 	if err != nil && err.Error() != "unauthorized" {
 		t.Fatal("alice improperly authorized to query")
 	}
@@ -368,4 +368,11 @@ func (m *mockCredentialStore) AA(username, password, perm string) bool {
 
 func mustNewMockCredentialStore() *mockCredentialStore {
 	return &mockCredentialStore{HasPermOK: true}
+}
+
+func makeCredentials(username, password string) *Credentials {
+	return &Credentials{
+		Username: username,
+		Password: password,
+	}
 }
