+++ http/service.go
@@ -15,7 +15,6 @@ import (
 	"net"
 	"net/http"
 	"net/http/pprof"
-	"net/url"
 	"os"
 	"runtime"
 	"strings"
@@ -23,6 +22,7 @@ import (
 	"time"
 
 	"github.com/rqlite/rqlite/auth"
+	"github.com/rqlite/rqlite/cluster"
 	"github.com/rqlite/rqlite/command"
 	"github.com/rqlite/rqlite/command/encoding"
 	"github.com/rqlite/rqlite/queue"
@@ -32,15 +32,6 @@ import (
 var (
 	// ErrLeaderNotFound is returned when a node cannot locate a leader
 	ErrLeaderNotFound = errors.New("leader not found")
-
-	// ErrUserInfoExists is returned when a join address already contains
-	// a username and a password.
-	ErrUserInfoExists = errors.New("userinfo exists")
-)
-
-const (
-	NO_USERNAME  = ""
-	NO_PASSSWORD = ""
 )
 
 type ResultsError interface {
@@ -98,10 +89,10 @@ type Cluster interface {
 	GetNodeAPIAddr(nodeAddr string, timeout time.Duration) (string, error)
 
 	// Execute performs an Execute Request on a remote node.
-	Execute(er *command.ExecuteRequest, nodeAddr string, username string, password string, timeout time.Duration) ([]*command.ExecuteResult, error)
+	Execute(er *command.ExecuteRequest, nodeAddr string, creds *cluster.Credentials, timeout time.Duration) ([]*command.ExecuteResult, error)
 
 	// Query performs an Query Request on a remote node.
-	Query(qr *command.QueryRequest, nodeAddr string, username string, password string, timeout time.Duration) ([]*command.QueryRows, error)
+	Query(qr *command.QueryRequest, nodeAddr string, creds *cluster.Credentials, timeout time.Duration) ([]*command.QueryRows, error)
 
 	// Stats returns stats on the Cluster.
 	Stats() (map[string]interface{}, error)
@@ -1107,7 +1098,7 @@ func (s *Service) execute(w http.ResponseWriter, r *http.Request) {
 			username = ""
 		}
 
-		results, resultsErr = s.cluster.Execute(er, addr, username, password, timeout)
+		results, resultsErr = s.cluster.Execute(er, addr, makeCredentials(username, password), timeout)
 		if resultsErr != nil && resultsErr.Error() == "unauthorized" {
 			http.Error(w, "remote execute not authorized", http.StatusUnauthorized)
 			return
@@ -1204,7 +1195,7 @@ func (s *Service) handleQuery(w http.ResponseWriter, r *http.Request) {
 		if !ok {
 			username = ""
 		}
-		results, resultsErr = s.cluster.Query(qr, addr, username, password, timeout)
+		results, resultsErr = s.cluster.Query(qr, addr, makeCredentials(username, password), timeout)
 		if resultsErr != nil && resultsErr.Error() == "unauthorized" {
 			http.Error(w, "remote query not authorized", http.StatusUnauthorized)
 			return
@@ -1345,7 +1336,7 @@ func (s *Service) runQueue() {
 								time.Sleep(retryDelay)
 								continue
 							}
-							_, err = s.cluster.Execute(er, addr, NO_USERNAME, NO_PASSSWORD, defaultTimeout)
+							_, err = s.cluster.Execute(er, addr, nil, defaultTimeout)
 							if err != nil {
 								s.logger.Printf("execute queue write failed for sequence number %d: %s",
 									req.SequenceNumber, err.Error())
@@ -1658,60 +1649,6 @@ func prettyEnabled(e bool) string {
 	return "disabled"
 }
 
-// NormalizeAddr ensures that the given URL has a HTTP protocol prefix.
-// If none is supplied, it prefixes the URL with "http://".
-func NormalizeAddr(addr string) string {
-	if !strings.HasPrefix(addr, "http://") && !strings.HasPrefix(addr, "https://") {
-		return fmt.Sprintf("http://%s", addr)
-	}
-	return addr
-}
-
-// EnsureHTTPS modifies the given URL, ensuring it is using the HTTPS protocol.
-func EnsureHTTPS(addr string) string {
-	if !strings.HasPrefix(addr, "http://") && !strings.HasPrefix(addr, "https://") {
-		return fmt.Sprintf("https://%s", addr)
-	}
-	return strings.Replace(addr, "http://", "https://", 1)
-}
-
-// CheckHTTPS returns true if the given URL uses HTTPS.
-func CheckHTTPS(addr string) bool {
-	return strings.HasPrefix(addr, "https://")
-}
-
-// AddBasicAuth adds username and password to the join address. If username is empty
-// joinAddr is returned unchanged. If joinAddr already contains a username, ErrUserInfoExists
-// is returned.
-func AddBasicAuth(joinAddr, username, password string) (string, error) {
-	if username == "" {
-		return joinAddr, nil
-	}
-
-	u, err := url.Parse(joinAddr)
-	if err != nil {
-		return "", err
-	}
-
-	if u.User != nil && u.User.Username() != "" {
-		return "", ErrUserInfoExists
-	}
-
-	u.User = url.UserPassword(username, password)
-	return u.String(), nil
-}
-
-// RemoveBasicAuth returns a copy of the given URL, with any basic auth information
-// removed.
-func RemoveBasicAuth(u string) string {
-	uu, err := url.Parse(u)
-	if err != nil {
-		return u
-	}
-	uu.User = nil
-	return uu.String()
-}
-
 // queryRequestFromStrings converts a slice of strings into a command.QueryRequest
 func executeRequestFromStrings(s []string, timings, tx bool) *command.ExecuteRequest {
 	stmts := make([]*command.Statement, len(s))
@@ -1752,4 +1689,11 @@ func queryRequestFromStrings(s []string, timings, tx bool) *command.QueryRequest
 // file. See https://www.sqlite.org/fileformat.html
 func validSQLiteFile(b []byte) bool {
 	return len(b) > 13 && string(b[0:13]) == "SQLite format"
+}
+
+func makeCredentials(username, password string) *cluster.Credentials {
+	return &cluster.Credentials{
+		Username: username,
+		Password: password,
+	}
 }
